+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - Kudo's to Alex Uta.
 - add performance example
   - add performance output files.
+- add support MCP3201 (experimental)
 - update examples (print lib version).
 - update readme.md
 - update library.properties